Define new TYPE and UPLINK vars for registration  (#27)
* Defines new TYPE and UPLINK vars for registration

The Autojoin API's register endpoint is going to start enforcing the presence
of these values. Today, all M-Lab managed machines has these values as part of
the Heartbeat registration process, and so should Autojoin machines.

* Adds TYPE and UPLINK vars to .env file for deploy.sh

* Temporarily sets DOCKER_TAG=latest for testing

* Revert "Temporarily sets DOCKER_TAG=latest for testing"

This reverts commit 032ab1cbfc41c20bb1ed239ec2d236cbf28641d4.

* Removes manual DOCKER_TAG var from cloudbuild.conf and deploy.sh

The value of this variable will now be determined dymaically at build time by
querying the repository directly for the latest tag.

* Makes DOCKER_TAG a user-editable var in Docker env file

We want to no longer user the "stable" tag. This prompts users to manually
verify the latest release of the autojoin repository and use that as the value
of the variable DOCKER_TAG.

* Set UPLINK=7g in deploy.sh

This value is more consistent with the actual max uplink rate for GCP VMs.
Also, updates USAGE, removing "<docker-tag>".

* Removes DOCKER_TAG variable from env file

Removes the DOCKER_TAG variable from the file env, and instead just statically
configures the latest released version of the autojoin repo (v0.2.11) in the
ndt-fullstack.yml Docker compose file.