Adds 9100 to list of ports sent to Autojoin API
Adding this port to the list of ports will cause Autojoin to return Prometheus
targets for node_exporter, which should cause the Prometheus instance in the
mlab-autojoin project to automatically start scraping node_exporter.

Also, even though 9100 is the default node_exporter port, explicitly set it in
the container specification so that it is more transparent in the compose file
whey we are sending port 9100 to Autojoin.